backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
GDAL @ file:///C:/Users/SungWook/Downloads/GDAL-3.4.3-cp311-cp311-win_amd64.whl
psycopg2==2.9.5
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
